feat<v0.5.0>: 重构index.jsindex.js
