+++ lib/utils.js
@@ -1,5 +1,6 @@
 const semver = require("semver");
-const request = require('request');
+const request = require("request");
+const chalk = require("chalk");
 
 /**
  * @description 判断node版本是否合适
@@ -8,10 +9,16 @@ const request = require('request');
  * @export
  * @param {*} wanted 想要的版本
  */
-export function checkNodeVersion(wanted = ">=12.18.3") {
+function checkNodeVersion(wanted = ">=12.18.3") {
 	// process.version 可以获取当前的 node 版本
 	if (!semver.satisfies(process.version, wanted)) {
-		console.log("Node版本不支持");
+		console.log(
+			"node 版本不支持您的版本" +
+				chalk.yellow(process.version) +
+				"，脚手架推荐使用" +
+				chalk.green("v12.18.3") +
+				"及以上版本。"
+		);
 		// 退出进程
 		process.exit(1);
 	}
@@ -35,3 +42,5 @@ function checkPackageVersion(url) {
 		});
 	});
 }
+
+module.exports.checkNodeVersion = checkNodeVersion;
